backend/pyproject.toml
[project]
name = "aimwallpaper-backend"
version = "0.1.0"
description = "AI-driven Wallpaper Engine wallpaper creation tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiofiles>=24.0.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "litellm>=1.40.0",
    "chromadb>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
